Move auth-machine to use service TraceProvider.
This PR moves the auth-machine service to use the service TraceProvider.
